requirements.txt
#libraries need to be installed
streamlit
streamlit_extras
google-genrativeai
python-dotenv
PyPDF2
Pillow